validatecommon.sh: 'git grep' for finding missing EOF newlines.
Processing time now 16.6 seconds. We're back in a usable range.
Yay!

This also de-duplicates the text file glob pattern.

An unintended side effect is, most warnings about not minimized
are gone, because testignore{} is no longer queried with a
complete list of all files.

As finding the right regexp took more than a day:

  -P   turns on Perl-compatilbe regular expressions,

  (?m) turns on multi-line mode, the whole file gets evaluated
       as ome single string,

  \N   matches everything but a newline,

  \z   matches end of the subject, which is in multi-line mode
       also end of the file ($ matches end of each line).