AdminProductsController::initProcess: init object
This partially reverts commit e563ae2e5aa13a15a06acbe3552621bfdfb15384.

In ps16 / early thirtybees version the contoller object was initialized
in the initProcess. This was removed as part of commit above, and it
caused small compatibility issue. There can be third party modules or
overrides that expects object model to be initialized.

For example, see https://forum.thirtybees.com/topic/4934-custom-product-fields-are-not-working-after-migration-to-tb/#comment-39825