New class ModuleUpdate.
This class aims to implement what's done in module 'tbupdater' so
far.

Module 'tbupdater' wants to be a self-sufficient application,
which makes it very complex and copies a lot of code from core.
Having module update code in a module makes it impossible to
update this module.

All this hassle is reason enough to abandon this module.