ModuleUpdate: implement updateModule().
Also downloadModuleArchive() and unpackModule() to serve this
method. This replaces updateModule() of module 'tbupdater'.

Strategy is slightly different than in the module: downloading and
unpacking is done in two seperate methods. This will allow to use
this code for updating a module from a directly uploaded archive
as well.

Also use it in the only place where the previous one was used.

Also remove some dead code.