PaymentModule: consider an unpaid order as valid.
This is often the case, e.g. with bank wire or invoice payments.

Actually, this small commit is the result of a larger discussion
about flaws in the system of order statuses. It looks much like
order statuses should be seen as a convenience for merchants, only,
but modules should change order states, only. Order statuses would
be computed on demand, then.

This particular commit solves a roadblock when dealing with oders
being valid, but not yet paid.