AdminModulesController: fix styling of the file upload panel.
This is not only code simplification, this panel was also visually
oversized before and the field label was misplaced. Structure is
now about what the Options Helper class does (looked up at
Shipping -> Preferences).

This is quite tricky. Inside a <div> of class 'input-group', it
matters for styling, which child is the first and the last one.
Accordingly, having a hidden element there messes up visual
appearance.