ModuleUpdate: don't fetch from the server for every module list.
This was a glitch introduced when code was copied from module
'tbupdater'. A double-check for an already existing module list
cache was removed, without fully compensating consequences.

This glitch caused a server fetch for almost every back office
page load, and even three of them when loading the module list
page. OUCH!